+++ block/all.js
@@ -2,3 +2,6 @@ export { SlackDividerBlock, createSlackDividerBlock } from './divider.js';
 export { SlackHeaderBlock, createSlackHeaderBlock } from './header.js';
 export { SlackImageBlock, createSlackImageBlock } from './image.js';
 export { SlackSectionBlock, createSlackSectionBlock } from './section.js';
+export { SlackActionsBlock, createSlackActionsBlock } from './actions.js';
+export { SlackContextBlock, createSlackContextBlock } from './context.js';
+export { SlackInputBlock, createSlackInputBlock } from './input.js';