+++ element/all.js
@@ -2,3 +2,10 @@ export { SlackButtonElement, createSlackButtonElement } from './button.js';
 export { SlackMarkdownElement, createSlackMarkdownElement } from './markdown.js';
 export { SlackPlainTextElement, createSlackPlainTextElement } from './plain-text.js';
 export { SLACK_DEFAULT, SLACK_PRIMARY, SLACK_DANGER } from './styles.js';
+
+// Inputs
+export {
+	SlackTextInputElement, createSlackTextInputElement, SlackEmailInputElement,
+	createSlackEmailInputElement, SlackURLInputElement, createSlackURLInputElement,
+	SlackNumberInputElement, createSlackNumberInputElement,
+} from './input/all.js';